+++ ChangeLog.md
@@ -3,7 +3,6 @@
 ## TODO
 
 - Migrate platform to version catalog
-- Add `ezgradle.logLevel` to modify the log level
 - Add Git properties to BuildConfig class
 - Add `outdatedDependencies` task for project ezgradle-bom to show outdated dependencies
 - Add `updateDependencies` task for project ezgradle-bom to update outdated dependencies automatically
@@ -13,6 +12,7 @@
 ### 2024.03.01
 
 - Feature: Experimental for version catalog
+- Feature: Add `ezgradle.logLevel` to modify the log level
 - Refactor: Generate BuildConfig.java instead of BuildConfig.kt
 
 ### 2024.02.02